Big fucking commit ! (Dijkstra Objet version with NetworkX)
As we decided together due to data structure problems; a new version of our project is done: a Object oriented version.
It's tested, and all you have to do is to use test.py to test it as well.
Now we have to develop the two other ways to calculate the shortest path:
- Interesting one
- With minimum flow